requirements.txt
paho-mqtt
influxdb-client
fastapi
uvicorn
